% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPBNfunctions.R
\name{llik_ZIPBN_j}
\alias{llik_ZIPBN_j}
\title{Function for evaluating log-likelihood of each observation for the j-th component of ZIPBN}
\usage{
llik_ZIPBN_j(x_j, logitPi, logLambda)
}
\arguments{
\item{x_j}{data for the j-th variable (node)}

\item{logitPi}{logit(pi)}

\item{logLambda}{log(lambda)}
}
\value{
log-likelihood of each observation for the j-th component of ZIPBN
}
\description{
Function for evaluating log-likelihood of each observation for the j-th component of ZIPBN
}
