\name{ZIPBN-package}
\alias{ZIPBN-package}
\alias{ZIPBN}
\docType{package}
\title{
\packageTitle{ZIPBN}
}
\description{
\packageDescription{ZIPBN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ZIPBN}
\packageIndices{ZIPBN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ZIPBN}

Maintainer: \packageMaintainer{ZIPBN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
