% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPBNfunctions.R
\name{mcmc_ZIPBN}
\alias{mcmc_ZIPBN}
\title{Implementation of MCMC sampler for ZIPBN models}
\usage{
mcmc_ZIPBN(x, starting, tuning, priors, n_samples = 5000)
}
\arguments{
\item{x}{a matrix containing data}

\item{starting}{a list of parameters' starting values for MCMC}

\item{tuning}{a list of precision values for Metropolis sampler Normal proposal distribution}

\item{priors}{a list of hyperparameter values for priors}

\item{n_samples}{the number of MCMC iterations}
}
\value{
MCMC samples from posterior distributions of ZIPBN models
}
\description{
Implementation of MCMC sampler for ZIPBN models
}
